Changes with nginx 1.2.4                                         25 Sep 2012
*) Bugfix: in the "limit_req" directive; the bug had appeared in 1.1.14.
Thanks to Charles Chen.

*) Bugfix: nginx could not be built by gcc 4.7 with -O2 optimization if
the --with-ipv6 option was used.

*) Bugfix: a segmentation fault might occur in a worker process if the
"map" directive was used with variables as values.

*) Bugfix: a segmentation fault might occur in a worker process if the
"geo" directive was used with the "ranges" parameter but without the
"default" parameter; the bug had appeared in 0.8.43.
Thanks to Zhen Chen and Weibin Yao.

*) Bugfix: in the -p command-line parameter handling.

*) Bugfix: in the mail proxy server.

*) Bugfix: of minor potential bugs.
Thanks to Coverity.

*) Bugfix: nginx/Windows could not be built with Visual Studio 2005
Express.
Thanks to HAYASHI Kentaro.